Deduplicate content from README.prod.md.
Right now, this is more nicely formatted than the README.prod.md,
but it's duplicative, and we're updating README.prod.md much
more frequently, which means server.html is out of date. Until
we've stabilized the division of Read the Docs/Zulip.org docs,
we'll only list requirements, the sha1sums of the Zulip tarballs,
and a few other minimal notes here, and point to README.prod.md
for installation instructions.

Fixes: zulip/zulip#668.